batik-ml-service/requirements.txt
# Modern Python 3.11+ Requirements
# Upgraded from old Python 2.7 + TF 1.6

# Core ML - Use latest compatible versions
tensorflow>=2.15.0
numpy>=1.24.0,<2.0.0
scipy>=1.11.0
scikit-learn>=1.3.0
h5py>=3.10.0

# Image Processing
opencv-python==4.8.1.78
Pillow==10.1.0
albumentations==1.3.1

# API
fastapi==0.109.0
uvicorn==0.25.0
python-multipart==0.0.6
pydantic==2.5.3

# Utilities
python-dotenv==1.0.0
tqdm==4.66.1
joblib==1.3.2

# Development
pytest==7.4.3
black==23.12.1
flake8==7.0.0

# Removed (outdated):
# - tensorflow-gpu (now just tensorflow with GPU auto-detect)
# - keras (now integrated in tensorflow)
# - tables==3.4.2 (not needed)
# - imutils==0.4.5 (use opencv directly)
# - opencv-contrib-python (use standard opencv)
